simle int parsing in asm
