+++ lib/utils.js
@@ -9,6 +9,7 @@ const NUMBYTES = constants.NUMBYTES, SHIFT32 = constants.SHIFT32
 const MAX_SAFE_HIGH = 0x1fffff
 
 exports.parseCBORint = function (ai, buf) {
+  console.log('parsecborint', ai, buf)
   var f, g
   switch (ai) {
     case NUMBYTES.ONE:
